Adapt to changes in terminal size
changelog:
- rerender the screen when the terminal size changes
