+++ src/utils.cpp
@@ -6,98 +6,6 @@
 
 #include "utils.h"
 
-utils::Ncurses::Ncurses() {
-    initscr();
-    raw();
-    noecho();
-    curs_set(0);
-}
-
-utils::Ncurses::~Ncurses() {
-    endwin();
-}
-
-utils::Ncurses::Screen::Screen(int height, int width, int y, int x) : lines_(height - 1), cols_(width - 1) {
-    window_ = newwin(height, width, y, x);
-    if (window_ == nullptr) throw std::runtime_error("Failed to create a window");
-    keypad(window_, true);
-}
-
-utils::Ncurses::Screen::~Screen() {
-    delwin(window_);
-}
-
-auto utils::Ncurses::Screen::getCharLower() -> char {
-    return static_cast<char>(std::tolower(wgetch(window_)));
-}
-
-void utils::Ncurses::Screen::putLineAt(const std::string &string, int y, int x) {
-    wmove(window_, y, 0);
-    wclrtoeol(window_);
-    wmove(window_, y, x);
-    waddstr(window_, string.c_str());
-    wrefresh(window_);
-}
-
-void utils::Ncurses::Screen::putLineFor(const std::string &string, int y, int x, std::chrono::seconds duration) {
-    putLineAt(string, y, x);
-    wrefresh(window_);
-    std::this_thread::sleep_for(duration);
-    wmove(window_, y, 0);
-    wclrtoeol(window_);
-    wrefresh(window_);
-}
-
-void utils::Ncurses::Screen::putLineWrapped(const std::string &string, int y, int x, int width) {
-    auto strLen{string.length()};
-    for (auto i = 0, l = 0; i < strLen; l++) {
-        auto charWrappedLine = string.substr(i, width);
-        auto lastCharIdx{charWrappedLine.length() - 1};
-        if (charWrappedLine[lastCharIdx] != '\n' && charWrappedLine[lastCharIdx] != ' ') {
-            auto lastSpace{charWrappedLine.find_last_of(' ')};
-            auto wordWrappedLine = charWrappedLine.substr(0, lastSpace);
-            i += static_cast<int>(wordWrappedLine.size());
-            putLineAt(wordWrappedLine, y + l, x);
-        } else {
-            putLineAt(charWrappedLine, y + l, x);
-            i += width;
-        }
-    }
-}
-
-
-auto
-utils::Ncurses::Screen::ask(const std::string &string, const std::string &validChars, unsigned int retries) -> char {
-    char ans;
-    while (retries--) {
-        putLineAt(string, 0, 0);
-        ans = getCharLower();
-        if (validChars.find(ans) == std::string::npos) {
-            putLineFor(std::string(" is not a valid answer").insert(0, 1, ans), lines_ / 2, cols_ / 2 - 11,
-                       std::chrono::seconds(1));
-            ans = '\0';
-        }
-    }
-    return ans;
-}
-
-void utils::Ncurses::Screen::clearScreen() {
-    wclear(window_);
-}
-
-auto utils::Ncurses::Screen::getLines() const -> int {
-    return lines_;
-}
-
-auto utils::Ncurses::Screen::getCols() const -> int {
-    return cols_;
-}
-
-utils::Ncurses::Screen::Screen(WINDOW *window) : lines_(LINES - 1), cols_(COLS - 1) {
-    window_ = window;
-    keypad(window_, true);
-}
-
 std::string utils::executeProcess(const std::string &path, const std::vector<const char *> &args) noexcept(false) {
     int fields[2];  // 0: read fd, 1: write fd
     pipe(fields);
